+++ app/api/chemotion/screen_api.rb
@@ -77,6 +77,7 @@ module Chemotion
         optional :result, type: String
         optional :description, type: Hash
         requires :wellplate_ids, type: Array
+        requires :research_plan_ids, type: Array
         requires :container, type: Hash
         optional :segments, type: Array, desc: 'Segments'
       end
@@ -121,6 +122,7 @@ module Chemotion
         optional :description, type: Hash
         optional :collection_id, type: Integer
         requires :wellplate_ids, type: Array
+        requires :research_plan_ids, type: Array
         requires :container, type: Hash
         optional :segments, type: Array, desc: 'Segments'
       end
@@ -131,7 +133,8 @@ module Chemotion
           requirements: params[:requirements],
           conditions: params[:conditions],
           result: params[:result],
-          description: params[:description]
+          description: params[:description],
+          research_plan_ids: params[:research_plan_ids]
         }
 
         screen = Screen.create(attributes)
@@ -145,19 +148,19 @@ module Chemotion
         recent_ols_term_update('chmo', kinds) if kinds&.length&.positive?
 
         collection = current_user.collections.where(id: params[:collection_id]).take
-        CollectionsScreen.create(screen: screen, collection: collection) if collection.present?
+        screen.collections << collection if collection.present?
 
         is_shared_collection = false
         unless collection.present?
           sync_collection = current_user.all_sync_in_collections_users.where(id: params[:collection_id]).take
           if sync_collection.present?
             is_shared_collection = true
-            CollectionsScreen.create(screen: screen, collection: Collection.find(sync_collection['collection_id']))
-            CollectionsScreen.create(screen: screen, collection: Collection.get_all_collection_for_user(sync_collection['shared_by_id']))
+            screen.collections << Collection.find(sync_collection['collection_id'])
+            screen.collections << Collection.get_all_collection_for_user(sync_collection['shared_by_id'])
           end
         end
 
-        CollectionsScreen.create(screen: screen, collection: Collection.get_all_collection_for_user(current_user.id)) unless is_shared_collection
+        screen.collections << Collection.get_all_collection_for_user(current_user.id) unless is_shared_collection
 
         params[:wellplate_ids].each do |id|
           ScreensWellplate.find_or_create_by(wellplate_id: id, screen_id: screen.id)