+++ app/packs/src/components/WellOverlay.js
@@ -1,134 +1,42 @@
 import React from 'react';
-import {
-  Button,
-  Popover,
-  Overlay,
-  ControlLabel,
-  FormGroup,
-  FormControl,
-  Col, InputGroup, ButtonGroup, Form,
-} from 'react-bootstrap';
+import { Button, Popover, Overlay, ControlLabel, FormGroup, FormControl, Col, InputGroup, ButtonGroup } from 'react-bootstrap';
 import SVG from 'react-inlinesvg';
-import UIStore from './stores/UIStore';
-import { wellplateShowSample } from './routesUtils';
+import PropTypes from 'prop-types';
 import Select from 'react-select';
 import { CirclePicker } from 'react-color';
+import { wellplateShowSample } from './routesUtils';
 
-const WellOverlay = ({show, well, placement, target, handleClose, removeSampleFromWell, handleWellLabel, handleColorPicker, selectedColor, saveColorCode}) => {
-  return (
-    <Overlay  rootClose
-              show={show}
-              target={target}
-              placement={placement}
-              onHide={() => handleClose()} >
-      <Popover title={title(handleClose)} id={'wellpop'+well.id}>
-        {content(well, removeSampleFromWell, handleWellLabel, handleColorPicker, selectedColor, saveColorCode)}
-      </Popover>
-    </Overlay>
-  );
-}
+const sampleName = (sample) => { /* eslint-disable camelcase */
+  if (sample) {
+    const { name, external_label, short_label } = sample;
+    const sampleNameLabel = `${name || ''} ${external_label || ''} ${short_label || ''}`;
+    if (sample.isNew) {
+      return sampleNameLabel;
+    }
+    return (
+      // TODO: FIXME: Uncaught ReferenceError: handleSampleClick is not defined
+      <a onClick={() => handleSampleClick(sample)} style={{ cursor: 'pointer' }}>
+        {sampleNameLabel}
+      </a>
+    );
+  }
+  return (<div />);
+}; /* eslint-enable */
 
-const content = (well, removeSampleFromWell, handleWellLabel, handleColorPicker, selectedColor, saveColorCode) => {
-  const { sample } = well;
-  const bcStyle = {
-    backgroundColor: selectedColor || well.color_code
-  };
-  const wellLabels = well.label ? well.label.split(',') : [];
-  const isDisable = () => wellLabels.some(item => item === 'Molecular structure');
-
-  const labels = [{
-    label: 'Name',
-    value: 'Name',
-    disabled: isDisable()
-  }, {
-    label: 'External label',
-    value: 'External label',
-    disabled: isDisable()
-  }, {
-    label: 'Molecular structure',
-    value: 'Molecular structure',
-    disabled: (wellLabels.some(item => item !== 'Molecular structure'))
-  }];
-
-  return(
-    <div style={{width: 220, height: 850}}>
-      {renderWellContent(well, removeSampleFromWell)}
-      <div>
-        <hr style={{marginTop: 28, marginBottom: 10}}/>
-        <Select
-          id="label"
-          name="label"
-          multi
-          options={labels}
-          value={well.label}
-          onChange={e => handleWellLabel(e)}
-        />
-        &nbsp;
-        <FormGroup>
-          <ControlLabel>Readout</ControlLabel>
-          <FormControl componentClass="textarea"
-            disabled={true}
-            value={well.readout || ''}
-            style={{height: 80}}
-          />
-        </FormGroup>
-        <FormGroup>
-          <ControlLabel>Imported Readout</ControlLabel>
-          <FormControl componentClass="textarea"
-            disabled={true}
-            value={sampleImportedReadout(sample) || ''}
-            style={{height: 80}}
-          />
-        </FormGroup>
-        <FormGroup style={{ top: '50px' }} controlId="colorInput">
-          <Col componentClass={ControlLabel} sm={3}>
-            Select Color
-          </Col>
-          <Col sm={9}>
-            <InputGroup>
-              <InputGroup.Addon style={bcStyle}></InputGroup.Addon>
-              <FormControl
-                type="text"
-                readOnly
-                value={selectedColor || well.color_code}
-              />
-            </InputGroup>
-          </Col>
-        </FormGroup>
-        <FormGroup controlId="formHorizontalPicker">
-          <Col sm={12}>
-            <CirclePicker width="132%" onChangeComplete={e => handleColorPicker(e)} />
-          </Col>
-        </FormGroup>
-        <ButtonGroup style={{ top: '10px', bottom: '10px' }}>
-          <Button style={{ left: '80px' }} onClick={saveColorCode}>Save</Button>
-        </ButtonGroup>
-      </div>
-    </div>
-  )
-}
-
-const title = (handleClose) => {
-  return(
-    <div>
-      Well Details
-      <span className='pull-right' style={{marginRight: -8, marginTop: -3}}>
-        <Button bsSize='xsmall' onClick={() => handleClose()}>
-          <i className="fa fa-times"></i>
-        </Button>
-      </span>
-    </div>
-  )
+const handleSampleClick = (sample) => {
+  const { params, uri } = Aviator.getCurrentRequest();
+  Aviator.navigate(`${uri}/sample/${sample.id}`, { silent: true });
+  wellplateShowSample({ params: { ...params, sampleID: sample.id } });
 }
 
 const renderWellContent = (well, removeSampleFromWell) => {
   const {sample} = well;
   let svg, moleculeName, removeButton = '';
-  const namesStyle= {textAlign: 'center', marginTop: 5};
+
   const svgContainerStyle = {
     borderRadius: '50%',
-    height: 200,
-    width: 200,
+    height: 190,
+    width: 190,
     border: '6px solid lightgray',
     textAlign: 'center',
     verticalAlign: 'middle',
@@ -150,42 +58,156 @@ const renderWellContent = (well, removeSampleFromWell) => {
       <div style={svgContainerStyle}>
         {svg}
       </div>
-      <div style={namesStyle}>
+      <div className="wellplate-overlay">
         {sampleName(sample)}<br/>
         {moleculeName}<br/>
       </div>
-      {removeButton}
+      <div>
+        {removeButton}
+      </div>
     </div>
   );
 }
 
-const sampleImportedReadout = (sample) => {
-  return sample
-    ? sample.imported_readout
-    : ''
+const sampleImportedReadout = sample => (sample ? sample.imported_readout : '');
+
+const content = (
+  well, readoutTitles, removeSampleFromWell, handleWellLabel, handleColorPicker,
+  selectedColor, saveColorCode
+) => {
+  const { sample, readouts } = well;
+  const bcStyle = {
+    backgroundColor: selectedColor || well.color_code
+  };
+  const wellLabels = well.label ? well.label.split(',') : [];
+  const isDisable = () => wellLabels.some(item => item === 'Molecular structure');
+
+  const labels = [{
+    label: 'Name',
+    value: 'Name',
+    disabled: isDisable()
+  }, {
+    label: 'External label',
+    value: 'External label',
+    disabled: isDisable()
+  }, {
+    label: 'Molecular structure',
+    value: 'Molecular structure',
+    disabled: (wellLabels.some(item => item !== 'Molecular structure'))
+  }];
+
+  return(
+    <div style={{ width: 220, height: 650 }}>
+      {renderWellContent(well, removeSampleFromWell)}
+      <div>
+        <Select
+          className="well-overlay-select"
+          id="label"
+          name="label"
+          multi
+          options={labels}
+          value={well.label}
+          onChange={e => handleWellLabel(e)}
+          style={{ top: '2px', bottom: '2px' }}
+        />
+        <FormGroup>
+          {readouts && readouts.map((readout, index) => (
+            <div key={`readout_${readout.id}`}>
+              <ControlLabel>{readoutTitles[index]}</ControlLabel>
+              <InputGroup>
+                <FormControl
+                  type="text"
+                  value={readout.value}
+                  disabled
+                  placeholder="Value"
+                />
+                <InputGroup.Addon disabled>{readout.unit}</InputGroup.Addon>
+              </InputGroup>
+            </div>
+          ))}
+        </FormGroup>
+        <FormGroup style={{ display: 'none' }}>
+          <ControlLabel>Imported Readout</ControlLabel>
+          <FormControl
+            componentClass="textarea"
+            disabled
+            value={sampleImportedReadout(sample) || ''}
+            style={{ height: 50 }}
+          />
+        </FormGroup>
+        <FormGroup>
+          <Col style={{ marginLeft: '-15px' }} class="row row-no-gutters" componentClass={ControlLabel} sm={3}>
+            Select&nbsp;Color
+          </Col>
+          <Col sm={9} style={{ marginLeft: '35px', width: '65%' }}>
+            <InputGroup>
+              <InputGroup.Addon style={bcStyle} />
+              <FormControl
+                class="input-sm"
+                type="text"
+                readOnly
+                value={selectedColor || well.color_code}
+              />
+            </InputGroup>
+          </Col>
+        </FormGroup>
+        <FormGroup controlId="formHorizontalPicker" style={{ marginTop: '60px' }}>
+          <CirclePicker circleSize={17} width="100%" onChangeComplete={e => handleColorPicker(e)} />
+        </FormGroup>
+        <ButtonGroup style={{ bottom: '5x' }}>
+          <Button style={{ left: '80px' }} onClick={saveColorCode}>Save</Button>
+        </ButtonGroup>
+      </div>
+    </div>
+  );
+};
+
+const title = (handleClose) => {
+  return(
+    <div>
+      Well Details
+      <span className='pull-right' style={{marginRight: -8, marginTop: -3}}>
+        <Button bsSize='xsmall' onClick={() => handleClose()}>
+          <i className="fa fa-times"></i>
+        </Button>
+      </span>
+    </div>
+  )
 }
 
-const sampleName = (sample) => {
-  if(sample) {
-    let {name, external_label, short_label} = sample
-    let sampleNameLabel = `${name || ""} ${external_label || ""} ${short_label || ""}`
-    if(sample.isNew) {
-      return sampleNameLabel;
-    } else {
-      return (
-        <a onClick={() => handleSampleClick(sample)} style={{cursor: 'pointer'}}>
-          {sampleNameLabel}
-        </a>
-      );
-    }
+const WellOverlay = ({
+  show, well, readoutTitles, placement, target, handleClose, removeSampleFromWell, handleWellLabel,
+  handleColorPicker, selectedColor, saveColorCode
+}) => (
+  <Overlay
+    rootClose
+    show={show}
+    target={target}
+    placement={placement}
+    style={{ position: 'sticky', top: 0, overflow: 'scroll' }}
+    onHide={() => handleClose()}
+  >
+    <Popover title={title(handleClose)} id={`wellpop${well.id}`}>
+      {content(
+        well, readoutTitles, removeSampleFromWell, handleWellLabel, handleColorPicker,
+        selectedColor, saveColorCode
+      )}
+    </Popover>
+  </Overlay>
+);
 
-  }
-}
-
-const handleSampleClick = (sample) => {
-  const { params, uri } = Aviator.getCurrentRequest();
-  Aviator.navigate(`${uri}/sample/${sample.id}`, { silent: true });
-  wellplateShowSample({ params: { ...params, sampleID: sample.id } });
-}
+WellOverlay.propTypes = {
+  show: PropTypes.bool.isRequired,
+  well: PropTypes.object.isRequired,
+  readoutTitles: PropTypes.array.isRequired,
+  placement: PropTypes.string.isRequired,
+  target: PropTypes.func.isRequired,
+  handleClose: PropTypes.func.isRequired,
+  removeSampleFromWell: PropTypes.func.isRequired,
+  handleWellLabel: PropTypes.func.isRequired,
+  handleColorPicker: PropTypes.func.isRequired,
+  selectedColor: PropTypes.func.isRequired,
+  saveColorCode: PropTypes.func.isRequired,
+};
 
 export default WellOverlay;