+++ app/packs/src/components/research_plan/ResearchPlanDetailsAttachments.js
@@ -1,12 +1,24 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
 import Dropzone from 'react-dropzone';
-import { FormGroup, Button, Row, Col, Tooltip, ControlLabel, ListGroup, ListGroupItem, OverlayTrigger } from 'react-bootstrap';
-import { last, findKey, values } from 'lodash';
 import EditorFetcher from '../fetchers/EditorFetcher';
+import ElementActions from '../actions/ElementActions';
 import ImageModal from '../common/ImageModal';
+import LoadingActions from '../actions/LoadingActions';
+import PropTypes from 'prop-types';
+import React, { Component } from 'react';
 import SpinnerPencilIcon from '../common/SpinnerPencilIcon';
-import { previewAttachmentImage } from './../utils/imageHelper';
+import {
+  Button, ButtonGroup,
+  Col, ControlLabel,
+  FormGroup,
+  Glyphicon,
+  ListGroup, ListGroupItem,
+  Overlay, OverlayTrigger,
+  Popover,
+  Row,
+  Tooltip
+} from 'react-bootstrap';
+import { last, findKey, values } from 'lodash';
+import AttachmentFetcher from '../fetchers/AttachmentFetcher';
 
 const editorTooltip = exts => <Tooltip id="editor_tooltip">Available extensions: {exts}</Tooltip>;
 const downloadTooltip = <Tooltip id="download_tooltip">Download attachment</Tooltip>;
@@ -15,23 +27,28 @@ const imageStyle = { position: 'absolute', width: 60, height: 60 };
 export default class ResearchPlanDetailsAttachments extends Component {
   constructor(props) {
     super(props);
+    this.importButtonRefs = [];
     const {
       attachments, onDrop, onDelete, onUndoDelete, onDownload, onEdit
     } = props;
     this.state = {
-      onDrop,
-      onDelete,
-      onUndoDelete,
-      onDownload,
-      onEdit,
       attachmentEditor: false,
       extension: null,
+      showImportConfirm: []
     };
     this.editorInitial = this.editorInitial.bind(this);
+    this.createAttachmentPreviews = this.createAttachmentPreviews.bind(this);
   }
 
   componentDidMount() {
     this.editorInitial();
+    this.createAttachmentPreviews();
+  }
+
+  componentDidUpdate(prevProps) {
+    if (this.props.attachments !== prevProps.attachments) {
+      this.createAttachmentPreviews();
+    }
   }
 
   editorInitial() {
@@ -58,7 +75,6 @@ export default class ResearchPlanDetailsAttachments extends Component {
   }
 
   handleEdit(attachment) {
-    const { onEdit } = this.state;
     const fileType = last(attachment.filename.split('.'));
     const docType = this.documentType(attachment.filename);
 
@@ -71,27 +87,51 @@ export default class ResearchPlanDetailsAttachments extends Component {
           attachment.aasm_state = 'oo_editing';
           attachment.updated_at = new Date();
 
-          onEdit(attachment);
+          this.props.onEdit(attachment);
         } else {
           alert('Unauthorized to edit this file.');
         }
       });
   }
 
-  renderRemoveAttachmentButton(attachment) {
-    const { onDelete } = this.state;
+  createAttachmentPreviews() {
+    const { attachments } = this.props;
+    attachments.map((attachment) => {
+      if (attachment.thumb) {
+        AttachmentFetcher.fetchThumbnail({ id: attachment.id }).then((result) => {
+          if (result != null) {
+            attachment.preview = `data:image/png;base64,${result}`;
+          }
+        });
+      }
+      else {
+        attachment.preview = '/images/wild_card/not_available.svg';
+      }
+      return attachment;
+    });
+  }
 
+  onImport(attachment) {
+    const researchPlanId = this.props.researchPlan.id;
+    LoadingActions.start();
+    ElementActions.importTableFromSpreadsheet(
+      researchPlanId,
+      attachment.id,
+      this.props.onAttachmentImportComplete
+    )
+    LoadingActions.stop();
+  }
+
+  renderRemoveAttachmentButton(attachment) {
     return (
-      <Button bsSize="xsmall" bsStyle="danger" className="button-right" onClick={() => onDelete(attachment)} disabled={this.props.readOnly}>
+      <Button bsSize="xsmall" bsStyle="danger" className="button-right" onClick={() => this.props.onDelete(attachment)} disabled={this.props.readOnly}>
         <i className="fa fa-trash-o" aria-hidden="true" />
       </Button>
     );
   }
 
   renderListGroupItem(attachment) {
-    const {
-      attachmentEditor, extension, onUndoDelete, onDownload
-    } = this.state;
+    const { attachmentEditor, extension } = this.state;
 
     const updateTime = new Date(attachment.updated_at);
     updateTime.setTime(updateTime.getTime() + (15 * 60 * 1000));
@@ -99,8 +139,6 @@ export default class ResearchPlanDetailsAttachments extends Component {
     const hasPop = false;
     const fetchNeeded = false;
     const fetchId = attachment.id;
-
-    const previewImg = previewAttachmentImage(attachment);
     const isEditing = attachment.aasm_state === 'oo_editing' && new Date().getTime() < updateTime;
 
     const docType = this.documentType(attachment.filename);
@@ -120,7 +158,7 @@ export default class ResearchPlanDetailsAttachments extends Component {
                 bsSize="xsmall"
                 bsStyle="danger"
                 className="button-right"
-                onClick={() => onUndoDelete(attachment)}
+                onClick={() => this.props.onUndoDelete(attachment)}
               >
                 <i className="fa fa-undo" aria-hidden="true" />
               </Button>
@@ -140,11 +178,11 @@ export default class ResearchPlanDetailsAttachments extends Component {
                   imageStyle={imageStyle}
                   hasPop={hasPop}
                   previewObject={{
-                    src: previewImg
+                    src: attachment.preview,
                   }}
                   popObject={{
                     title: attachment.filename,
-                    src: previewImg,
+                    src: attachment.preview,
                     fetchNeeded,
                     fetchId
                   }}
@@ -162,7 +200,7 @@ export default class ResearchPlanDetailsAttachments extends Component {
                 bsSize="xsmall"
                 className="button-right"
                 bsStyle="primary"
-                onClick={() => onDownload(attachment)}
+                onClick={() => this.props.onDownload(attachment)}
               >
                 <i className="fa fa-download" aria-hidden="true" />
               </Button>
@@ -179,6 +217,7 @@ export default class ResearchPlanDetailsAttachments extends Component {
                 <SpinnerPencilIcon spinningLock={!attachmentEditor || isEditing} />
               </Button>
             </OverlayTrigger>
+            {this.renderImportAttachmentButton(attachment)}
           </Col>
         </Row>
       </div>
@@ -194,30 +233,112 @@ export default class ResearchPlanDetailsAttachments extends Component {
             <ListGroupItem key={attachment.id}>
               {this.renderListGroupItem(attachment)}
             </ListGroupItem>
-            ))}
+          ))}
         </ListGroup>
       );
     }
     return (
       <div>
-        There are currently no Datasets.<br />
+        There are currently no attachments.<br />
       </div>
     );
   }
 
   renderDropzone() {
-    const { onDrop } = this.state;
 
     return (
-      <div className={`research-plan-dropzone-${this.props.readOnly ? 'disable' : 'enable'}`}>
-        <Dropzone
-          onDrop={files => onDrop(files)}
-          className="zone"
-        >
+      <Dropzone
+        onDrop={files => this.props.onDrop(files)}
+        className={`research-plan-dropzone-${this.props.readOnly ? 'disable' : 'enable'}`}
+      >
+        <div className="zone">
           Drop Files, or Click to Select.
-        </Dropzone>
-      </div>
+        </div>
+      </Dropzone>
     );
+  }
+
+  renderImportAttachmentButton(attachment) {
+    const show = this.state.showImportConfirm[attachment.id];
+    // TODO: import disabled when?
+    const importDisabled = this.props.researchPlan.changed;
+    const extension = last(attachment.filename.split('.'));
+
+    const importTooltip = importDisabled ?
+      <Tooltip id="import_tooltip">Research Plan must be saved before import</Tooltip> :
+      <Tooltip id="import_tooltip">Import spreadsheet as research plan table</Tooltip>;
+
+    const confirmTooltip = (
+      <Tooltip placement="bottom" className="in" id="tooltip-bottom">
+        Import data from Spreadsheet?<br />
+        <ButtonGroup>
+          <Button
+            bsStyle="success"
+            bsSize="xsmall"
+            onClick={() => this.confirmAttachmentImport(attachment)}
+          >
+            Yes
+          </Button>
+          <Button
+            bsStyle="warning"
+            bsSize="xsmall"
+            onClick={() => this.hideImportConfirm(attachment.id)}
+          >
+            No
+          </Button>
+        </ButtonGroup>
+      </Tooltip>
+    );
+
+    if (extension === 'xlsx') {
+      return (
+        <div>
+          <OverlayTrigger placement="top" overlay={importTooltip} >
+            <div style={{ float: 'right' }}>
+              <Button
+                bsSize="xsmall"
+                bsStyle="success"
+                className="button-right"
+                disabled={importDisabled}
+                ref={(ref) => { this.importButtonRefs[attachment.id] = ref; }}
+                style={importDisabled ? { pointerEvents: 'none' } : {}}
+                onClick={() => this.showImportConfirm(attachment.id)}
+              >
+                <Glyphicon glyph="import" />
+              </Button>
+            </div>
+          </OverlayTrigger>
+          <Overlay
+            show={show}
+            placement="bottom"
+            rootClose
+            onHide={() => this.hideImportConfirm(attachment.id)}
+            target={this.importButtonRefs[attachment.id]}
+          >
+            {confirmTooltip}
+          </Overlay>
+
+        </div>
+      );
+    }
+    return true;
+  }
+
+  showImportConfirm(attachmentId) {
+    const { showImportConfirm } = this.state;
+    showImportConfirm[attachmentId] = true;
+    this.setState({ showImportConfirm });
+  }
+
+  hideImportConfirm(attachmentId) {
+    const { showImportConfirm } = this.state;
+    showImportConfirm[attachmentId] = false;
+    this.setState({ showImportConfirm });
+  }
+
+  confirmAttachmentImport(attachment) {
+    this.onImport(attachment);
+    this.hideImportConfirm(attachment.id);
   }
 
   render() {
@@ -236,15 +357,18 @@ export default class ResearchPlanDetailsAttachments extends Component {
 }
 
 ResearchPlanDetailsAttachments.propTypes = {
+  researchPlan: PropTypes.object.isRequired,
   attachments: PropTypes.array,
   onDrop: PropTypes.func.isRequired,
   onDelete: PropTypes.func.isRequired,
   onUndoDelete: PropTypes.func.isRequired,
   onDownload: PropTypes.func.isRequired,
+  onAttachmentImportComplete: PropTypes,
   onEdit: PropTypes.func.isRequired,
   readOnly: PropTypes.bool.isRequired
 };
 
 ResearchPlanDetailsAttachments.defaultProps = {
-  attachments: []
+  attachments: [],
+  onAttachmentImportComplete: () => {}
 };