+++ spec/factories/users.rb
@@ -1,9 +1,15 @@
+# frozen_string_literal: true
+
 FactoryBot.define do
   factory :user do
     sequence(:email) { |n| "foobar#{n}@bar.de" }
     first_name { 'first_name' }
     last_name { 'last_name' }
-    sequence(:name_abbreviation) { [*'A'..'Z'].sample + SecureRandom.alphanumeric(2) }
+    sequence(:name_abbreviation) do |n|
+      ltrs = [*'a'..'z', *'A'..'Z'] # 52 alphabetic characters
+      chrs = [*0..9] + ltrs         # 62 alphanumeric characters
+      "#{ltrs[n / 52 / 62 % 62]}#{chrs[n / 62 % 62]}#{chrs[n % 62]}"
+    end
     password { 'testtest' }
     password_confirmation { 'testtest' }
     counters do
@@ -18,13 +24,13 @@ FactoryBot.define do
       profile = user.profile
       data = profile&.data
       unless data.nil?
-        data.merge!(layout: {
+        data[:layout] = {
           'sample' => 1,
           'reaction' => 2,
           'wellplate' => 3,
           'screen' => 4,
           'research_plan' => 5
-        })
+        }
         profile.update_columns(data: data)
       end
     end